+++ synth/SinOsc.h
@@ -19,12 +19,12 @@
 
 class SinOsc : public Gen {
 private:
-    static double *table;
-    double phaseacc;
-    double phase;
-    double freq; // actually freq/samprate
-    double pmamount; // amount of phase mod
-    double fmamount; // amount of freq mod
+    static float *table;
+    float phaseacc;
+    float phase;
+    float freq; // actually freq/samprate
+    float pmamount; // amount of phase mod
+    float fmamount; // amount of freq mod
     bool fixed; // fixed frequency, or multiple of key frequency
 public:
     
@@ -35,14 +35,14 @@ public:
         addin("fm",SINOSC_FM);
         
         fixed=false;
-        phaseacc=0;
-        pmamount=0.1;
-        fmamount=0.1;
+        phaseacc=0.0f;
+        pmamount=0.1f;
+        fmamount=0.1f;
         setFrequency(1);
         if(!table){
-            table = new double[SINETABLESIZE];
+            table = new float[SINETABLESIZE];
             for(int i=0;i<SINETABLESIZE;i++){
-                double x = (i*2*M_PI)/(double)SINETABLESIZE;
+                float x = (i*2.0f*M_PI)/(float)SINETABLESIZE;
                 table[i]=sin(x);
             }
         }
@@ -61,35 +61,35 @@ public:
     /// to some multiple of key (i.e. note
     /// played) frequency (this is the default).
     
-    void setFrequency(double f){
+    void setFrequency(float f){
         freq = f;
     }
     
-    void setPM(double m){
-        pmamount = m*0.01;
+    void setPM(float m){
+        pmamount = m*0.01f;
     }
     
-    void setFM(double m){
-        fmamount = m*0.01; // should be really tiny
+    void setFM(float m){
+        fmamount = m*0.01f; // should be really tiny
     }
     
     virtual void update(int nframes){
-        double *amp = ins[SINOSC_AMP];
-        double *pm = ins[SINOSC_PM];
-        double *fm = ins[SINOSC_FM];
+        float *amp = ins[SINOSC_AMP];
+        float *pm = ins[SINOSC_PM];
+        float *fm = ins[SINOSC_FM];
         
-        extern double samprate,keyFreq;
-        double f = freq;
+        extern float samprate,keyFreq;
+        float f = freq;
         if(!fixed)f*=keyFreq;
-        double step = f/samprate;
+        float step = f/samprate;
         
         for(int i=0;i<nframes;i++){
-            double istep = step + (fm?fm[i]*fmamount:0.0);
+            float istep = step + (fm?fm[i]*fmamount:0.0f);
             phaseacc += istep;
-            if(phaseacc>1.0)phaseacc -= 1.0;
-            if(phaseacc>1.0)phaseacc -= 1.0;
+            if(phaseacc>1.0f)phaseacc -= 1.0f;
+            if(phaseacc>1.0f)phaseacc -= 1.0f;
             // add 1 to avoid it going negative when we add the PM
-            double iphase = 1.0+phaseacc + (pm?pm[i]*pmamount:0.0);
+            float iphase = 1.0+phaseacc + (pm?pm[i]*pmamount:0.0);
             int x = iphase*SINETABLESIZE;
             x %= SINETABLESIZE;
             out[i]=table[x] * (amp?amp[i]:1.0);
